Switch from trying to expose the entire type signature to using the user_defined type
