+++ datafusion/ffi/src/util.rs
@@ -15,6 +15,11 @@
 // specific language governing permissions and limitations
 // under the License.
 
+use abi_stable::std_types::RVec;
+use arrow::{datatypes::DataType, ffi::FFI_ArrowSchema};
+
+use crate::arrow_wrappers::WrappedSchema;
+
 /// This macro is a helpful conversion utility to conver from an abi_stable::RResult to a
 /// DataFusion result.
 #[macro_export]
@@ -57,6 +62,27 @@ macro_rules! rresult_return {
             }
         }
     };
+}
+
+pub fn vec_datatype_to_rvec_wrapped(
+    data_types: &[DataType],
+) -> Result<RVec<WrappedSchema>, arrow::error::ArrowError> {
+    Ok(data_types
+        .iter()
+        .map(FFI_ArrowSchema::try_from)
+        .collect::<Result<Vec<_>, arrow::error::ArrowError>>()?
+        .into_iter()
+        .map(WrappedSchema)
+        .collect())
+}
+
+pub fn rvec_wrapped_to_vec_datatype(
+    data_types: &RVec<WrappedSchema>,
+) -> Result<Vec<DataType>, arrow::error::ArrowError> {
+    data_types
+        .iter()
+        .map(|d| DataType::try_from(&d.0))
+        .collect()
 }
 
 #[cfg(test)]
